requirements.txt
# Core dependencies
pyspark==3.5.0
kafka-python==2.0.2
cassandra-driver==3.28.0
redis==5.0.1
elasticsearch==8.11.0

# API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2
pyarrow==14.0.1
avro-python3==1.10.2

# Machine Learning
scikit-learn==1.3.2
tensorflow==2.15.0
xgboost==2.0.2
prophet==1.1.5

# Streaming
flink==1.18.0
pyflink==1.18.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0

# Data Generation & Faker
faker==20.1.0
mimesis==12.1.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0
configparser==6.0.0

# HTTP & API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0

# Monitoring
prometheus-client==0.19.0
py-zipkin==1.2.0

# Logging
loguru==0.7.2
python-json-logger==2.0.7

# Utilities
click==8.1.7
tqdm==4.66.1
schedule==1.2.0
python-dateutil==2.8.2
pytz==2023.3

# Geospatial
geopy==2.4.1
folium==0.15.1
geopandas==0.14.1

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
ipython==8.18.1
jupyter==1.0.0

# Docker SDK
docker==7.0.0

# Data Quality
great-expectations==0.18.8

# API Documentation
fastapi-swagger==0.2.2
